Extract ScriptRunner fixtures
