+++ tests/gui/measure_script/test_script_runner.py
@@ -10,26 +10,6 @@ from finesse.config import STEPPER_MOTOR_TOPIC
 from finesse.gui.measure_script.script import Script, ScriptRunner, _poll_em27_status
 
 
-@pytest.fixture
-def runner():
-    """Fixture for a ScriptRunner in its initial state."""
-    script = Script(Path(), 1, ({"angle": 0.0, "measurements": 3},))
-    runner = ScriptRunner(script)
-    runner._check_status_timer = MagicMock()
-    return runner
-
-
-@pytest.fixture
-def runner_measuring(
-    runner: ScriptRunner, subscribe_mock: MagicMock, sendmsg_mock: MagicMock
-) -> ScriptRunner:
-    """Fixture for a ScriptRunner in a measuring state."""
-    runner.start_moving()
-    runner.start_measuring()
-    sendmsg_mock.reset_mock()
-    return runner
-
-
 @patch("finesse.gui.measure_script.script.QTimer")
 def test_init(
     timer_mock: Mock, subscribe_mock: MagicMock, sendmsg_mock: MagicMock
